#7
ViewControllerからプログラムファイルを分割
